pyproject.toml
[tool.poetry]
name = "auth_sprint_1"
version = "0.1.0"
description = ""
authors = ["Skirda Matvey <skirdamatvei@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "2.1.2"
Flask-SQLAlchemy = "2.5.1"
connexion = {version = "2.13.1", extras = ["swagger-ui"]}
gevent = "21.12.0"
Werkzeug = "2.1.2"
requests = "2.23.0"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
Authlib = "0.14.1"
pyOpenSSL = "22.0.0"
SQLAlchemy = "1.4.36"
psycopg2-binary = "2.9.3"
pydantic = {version = "1.9.1", extras = ["dotenv"]}
Flask-JWT-Extended = "4.4.0"
passlib = "1.7.1"
Flask-Migrate = "3.1.0"
typer = "0.4.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
